Add files via upload
This dataset served as the foundation for the June 2013 release of the International Food Security Assessment, 1983-2013. For 76 low- and middle-income nations, this yearly ERS report forecasts food supply and access over a 10-year period. The dataset comprises annual country-level statistics on grain and root and tuber crops (combined as R&T in the documentation tables), food aid, the total value of imports and exports, gross domestic product, and population obtained from a number of sources. 

I took a look at the data and explored it a bit before I began the analysis:
•	We import packages we will be using for data analysis: Pandas, Numpy, Matplotlib and Seaborn.
•	We then read our data sets from the excel file we renamed (Data).

The food production of grain, cereals and root crops ‘Yield’, is non-linear across the years, there was an improvement in food production yield around 2012. The economy of Nigeria looks to be on the rise from 1983 to 2013 in our data. Unlike food yield, the population of Nigeria has been on a steady increase since 1983. Another factor worth considering from our exploratory data analysis (EDA) is the exportation and importation quantity of the country. Importation looks to be one significant influence on the food security of the country, Nigeria imports six times what they export. With all these factors that affect food security in Nigeria in place, the data shows the quantity of food available to be on a steady rise even though there was a drastic fall in 2002. From our two Models, we can predict the food availability in Nigeria in the next few years to be between 15,123.52 at the least bit and 20084.02 at the maximum. The two models suggest food availability will rise steadily assuming all factors remain the same.